add Tabs for settings
